move decam_depth view to oldviews
